+++ ormar/models/mixins/__init__.py
@@ -1,0 +1,5 @@
+from ormar.models.mixins.alias_mixin import AliasMixin
+from ormar.models.mixins.merge_mixin import MergeModelMixin
+from ormar.models.mixins.prefetch_mixin import PrefetchQueryMixin
+
+__all__ = ["MergeModelMixin", "AliasMixin", "PrefetchQueryMixin"]
