+++ ormar/queryset/clause.py
@@ -137,11 +137,9 @@ class QueryClause:
             if issubclass(model_cls.Meta.model_fields[part], ManyToManyField):
                 through_field = model_cls.Meta.model_fields[part]
                 previous_model = through_field.through
-                part2 = model_cls.resolve_relation_name(
-                    previous_model, through_field.to, explicit_multi=True
-                )
+                part2 = through_field.default_target_field_name()  # type: ignore
             manager = model_cls.Meta.alias_manager
-            table_prefix = manager.resolve_relation_join(previous_model, part2)
+            table_prefix = manager.resolve_relation_alias(previous_model, part2)
             model_cls = model_cls.Meta.model_fields[part].to
             previous_model = model_cls
         return select_related, table_prefix, model_cls
